@@ -16,12 +16,25 @@
         }
     },
     "require": {
-        "php": ">=7.2.0"
+        "php": ">=7.4.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "~6.2.0",
-        "squizlabs/php_codesniffer": "^3.0",
-        "friendsofphp/php-cs-fixer": "^2.5",
-        "satooshi/php-coveralls": "^1.0"
+        "friendsofphp/php-cs-fixer": "^2.16",
+        "phpunit/phpunit": "^9.1",
+        "squizlabs/php_codesniffer": "^3.5"
+    },
+    "scripts": {
+        "ci": [
+            "@lint",
+            "@test"
+        ],
+        "lint": [
+            "phpcbf",
+            "phpcs",
+            "php-cs-fixer fix"
+        ],
+        "test": [
+            "phpunit --colors --coverage-html ./coverage"
+        ]
     }
 }
